refactored song related stuff into its own subpackage
