+++ app/src/main/java/com/hich2000/tagcapella/songs/SongViewModel.kt
@@ -1,4 +1,4 @@
-package com.hich2000.tagcapella.music_player
+package com.hich2000.tagcapella.songs
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
